# SqlString - improved speed