--- a/lib/sql_string.js
+++ b/lib/sql_string.js
@@ -20,13 +20,13 @@ SqlString.escape = function(val, stringifyObjects, timeZone) {
   if (val instanceof Date) {
     val = SqlString.dateToString(val, timeZone || "Z");
   }
+  
+   if (Array.isArray(val)) {
+    return '(' + SqlString.arrayToList(val, timeZone) + ')';
+  }
 
   if (Buffer.isBuffer(val)) {
     return SqlString.bufferToString(val);
-  }
-
-  if (Array.isArray(val)) {
-    return SqlString.arrayToList(val, timeZone);
   }
 
   if (typeof val === 'object') {
@@ -52,10 +52,14 @@ SqlString.escape = function(val, stringifyObjects, timeZone) {
 };
 
 SqlString.arrayToList = function(array, timeZone) {
-  return array.map(function(v) {
-    if (Array.isArray(v)) return '(' + SqlString.arrayToList(v) + ')';
-    return SqlString.escape(v, true, timeZone);
-  }).join(', ');
+ var arrLength = array.length,
+     values = [];
+  
+  for (var i = 0; i < arrLength; i++) {
+    values.push(SqlString.escape(array[i], true, timeZone));
+  }
+  
+  return values.join(', ');
 };
 
 SqlString.format = function(sql, values, timeZone) {