@sidorares 2015-03-11
to do for me: rebase this and also look at https://github.com/felixge/node-mysql/commit/f39856c8b1edeb275a0daaccafe6b62f35e971a8 diff