@jtwebman 2016-01-20
Also @Mattijah I looked over this and you are not handling the nested array's in your optimization. That might be why is is faster.