@Mattijah 2016-01-21
@jtwebman good you mentioned that as it seems like the current solution is not work correctly.

**Instead of:**
```
INSERT INTO table VALUES ('a' 'b', 'c'), ('d', 'e', 'f');
```

**the output is:**
```
INSERT INTO table VALUES (('a' 'b', 'c'), ('d', 'e', 'f'));
```

I don't know how could I miss that, anyway... I am not sure if there is any point for me to spent time fixing this issue as this PR is here seven months not bothering many people.