@sidorares 2016-01-21
Hi @Mattijah - sorry it was low on my priority because 1) its perf only with no test 2) it's a change of sqlstring, and because most of the code from sqlstring and pool is taken from node-mysql but not shared properly ( as npm module ) I'm hesitant to do any changes there

Overall if there is performance benefit which is visible on a top scale ( e.i it's actually a hot path that contributes visible amount of time when measured connection-query x n-disconnect ) I'm really happy to see any experiments and PRs!