@Mattijah 2016-02-01
Hi @sidorares - Yeah, I understand. I just noticed now that the performance of `SqlString` was already improved here https://github.com/felixge/node-mysql/pull/1331. Have you seen it?