@sidorares 2016-02-01
yes, I'm keen to pull your change ( or that commit ). It looks very similar to what you've done, right?