@sidorares 2016-02-01
actually @Mattijah  if you can find time to rebase your changes and/or incorporate felixge/node-mysql#1331 and test with this library that would be really great