@sidorares 2016-06-09
@Mattijah I'm really sorry this never landed. We are using now sqlstring package shared between node-mysql and node-mysql2 - see https://github.com/mysqljs/sqlstring